워치로 부터 받아온 심박수 userInfo 에 저장 // 에러 해결 ERROR: Publishing changes from background threads is not allowed; make sure to publish values from the main thread (via operators like receive(on:)) on model updates.
- ERROR: Publishing changes from background threads is not allowed; make sure to publish values from the main thread (via operators like receive(on:)) on model updates.
-> DispatchQueue.main.async { [self] in } 사용하여 오류 해결

+++ watchConDemo/watchConDemo/Models/UserInfo.swift
@@ -18,7 +18,7 @@ class UserInfo {
     var accelerationY: [Double]?
     var accelerationZ: [Double]?
     var accelerationAvr: [Double]?
-    var heartRate: [Double]?
+    var heartRate: [Int]?
     var heartRateAvr: [Double]?
     var currentTime: [TimeInterval]?
